Moving `create_ruleset_from_lens` to entities module
